Modify tokens used in "sync-with-upstream" workflow
This change modifies the tokens used by the "sync-with-upstream"
workflow.

Now we use the token that has admin access when we checkout the
repository initially; this way the "sync-with-upstream" action has admin
access when performing git commands. This specifically is required when
the action attempts to git-push to the "sync-with-upstream" branch, in
order to overcome the usual branch protections.

Further, we don't use the token with admin access via the GITHUB_TOKEN
environment varable; this way the pull request that is opened by the
"sync-with-upstream" action is "owned" by the usual "github-actions"
user, rather than the user associated with the admin token. While this
isn't required, it makes it more consistent with PRs created by other
actions, as those PRs are owned by the "github-actions" user.